+++ lib/directives/index.dart
@@ -5,6 +5,7 @@ export 'tabs.dart';
 export 'highlightjs.dart';
 export 'clipboard.dart';
 export 'showdownjs.dart';
+export 'tippy.dart';
 
 import 'poppable.dart';
 import 'collapsible.dart';
@@ -13,6 +14,7 @@ import 'tabs.dart';
 import 'highlightjs.dart';
 import 'clipboard.dart';
 import 'showdownjs.dart';
+import 'tippy.dart';
 
 const ngAdminDirectives = [
   PoppableDirective,
@@ -24,5 +26,6 @@ const ngAdminDirectives = [
   WTabsDirective,
   HljsDirective,
   WClipboardDirective,
-  ShowdownDirective
+  ShowdownDirective,
+  TippyDirective
 ];